requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
celery>=5.3.6
redis>=5.0.1
bcrypt>=4.1.2
cryptography>=42.0.0
google-genai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
